mcp_server/requirements.txt
mcp[cli]>=1.0.0
uvicorn>=0.29.0
starlette>=0.37.0
psycopg2-binary>=2.9.9
